support exit on halt
when running runnit as process 1 inside a Docker container, it is
necessary for runit to actually *exit* after completing stage 3 in order
for Docker to clean up the container.

This patch adds support such that running "runit-init x" will cause
runit to exit rather than trying to halt or reboot the system.